split out endpoint parsing
