+++ CMakeLists.txt
@@ -2661,6 +2661,7 @@ add_library(grpc
   src/core/xds/grpc/xds_common_types.cc
   src/core/xds/grpc/xds_common_types_parser.cc
   src/core/xds/grpc/xds_endpoint.cc
+  src/core/xds/grpc/xds_endpoint_parser.cc
   src/core/xds/grpc/xds_health_status.cc
   src/core/xds/grpc/xds_http_fault_filter.cc
   src/core/xds/grpc/xds_http_filter_registry.cc
