+++ build_autogenerated.yaml
@@ -1232,6 +1232,7 @@ libs:
   - src/core/xds/grpc/xds_common_types.h
   - src/core/xds/grpc/xds_common_types_parser.h
   - src/core/xds/grpc/xds_endpoint.h
+  - src/core/xds/grpc/xds_endpoint_parser.h
   - src/core/xds/grpc/xds_health_status.h
   - src/core/xds/grpc/xds_http_fault_filter.h
   - src/core/xds/grpc/xds_http_filter.h
@@ -2043,6 +2044,7 @@ libs:
   - src/core/xds/grpc/xds_common_types.cc
   - src/core/xds/grpc/xds_common_types_parser.cc
   - src/core/xds/grpc/xds_endpoint.cc
+  - src/core/xds/grpc/xds_endpoint_parser.cc
   - src/core/xds/grpc/xds_health_status.cc
   - src/core/xds/grpc/xds_http_fault_filter.cc
   - src/core/xds/grpc/xds_http_filter_registry.cc
