+++ config.w32
@@ -836,6 +836,7 @@ if (PHP_GRPC != "no") {
     "src\\core\\xds\\grpc\\xds_common_types.cc " +
     "src\\core\\xds\\grpc\\xds_common_types_parser.cc " +
     "src\\core\\xds\\grpc\\xds_endpoint.cc " +
+    "src\\core\\xds\\grpc\\xds_endpoint_parser.cc " +
     "src\\core\\xds\\grpc\\xds_health_status.cc " +
     "src\\core\\xds\\grpc\\xds_http_fault_filter.cc " +
     "src\\core\\xds\\grpc\\xds_http_filter_registry.cc " +
