+++ gRPC-C++.podspec
@@ -1341,6 +1341,7 @@ Pod::Spec.new do |s|
                       'src/core/xds/grpc/xds_common_types_parser.h',
                       'src/core/xds/grpc/xds_enabled_server.h',
                       'src/core/xds/grpc/xds_endpoint.h',
+                      'src/core/xds/grpc/xds_endpoint_parser.h',
                       'src/core/xds/grpc/xds_health_status.h',
                       'src/core/xds/grpc/xds_http_fault_filter.h',
                       'src/core/xds/grpc/xds_http_filter.h',
@@ -2624,6 +2625,7 @@ Pod::Spec.new do |s|
                               'src/core/xds/grpc/xds_common_types_parser.h',
                               'src/core/xds/grpc/xds_enabled_server.h',
                               'src/core/xds/grpc/xds_endpoint.h',
+                              'src/core/xds/grpc/xds_endpoint_parser.h',
                               'src/core/xds/grpc/xds_health_status.h',
                               'src/core/xds/grpc/xds_http_fault_filter.h',
                               'src/core/xds/grpc/xds_http_filter.h',
