+++ gRPC-Core.podspec
@@ -2100,6 +2100,8 @@ Pod::Spec.new do |s|
                       'src/core/xds/grpc/xds_common_types_parser.h',
                       'src/core/xds/grpc/xds_endpoint.cc',
                       'src/core/xds/grpc/xds_endpoint.h',
+                      'src/core/xds/grpc/xds_endpoint_parser.cc',
+                      'src/core/xds/grpc/xds_endpoint_parser.h',
                       'src/core/xds/grpc/xds_health_status.cc',
                       'src/core/xds/grpc/xds_health_status.h',
                       'src/core/xds/grpc/xds_http_fault_filter.cc',
@@ -3403,6 +3405,7 @@ Pod::Spec.new do |s|
                               'src/core/xds/grpc/xds_common_types.h',
                               'src/core/xds/grpc/xds_common_types_parser.h',
                               'src/core/xds/grpc/xds_endpoint.h',
+                              'src/core/xds/grpc/xds_endpoint_parser.h',
                               'src/core/xds/grpc/xds_health_status.h',
                               'src/core/xds/grpc/xds_http_fault_filter.h',
                               'src/core/xds/grpc/xds_http_filter.h',
