+++ tools/doxygen/Doxyfile.c++.internal
@@ -2992,6 +2992,8 @@ src/core/xds/grpc/xds_common_types_parser.h \
 src/core/xds/grpc/xds_enabled_server.h \
 src/core/xds/grpc/xds_endpoint.cc \
 src/core/xds/grpc/xds_endpoint.h \
+src/core/xds/grpc/xds_endpoint_parser.cc \
+src/core/xds/grpc/xds_endpoint_parser.h \
 src/core/xds/grpc/xds_health_status.cc \
 src/core/xds/grpc/xds_health_status.h \
 src/core/xds/grpc/xds_http_fault_filter.cc \
